pyproject.toml
[project]
name = "therecruiter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyqt5>=5.15.9",
    "pyqt5-tools>=5.15.9.3.3",
    "mysql-connector-python>=8.0.0",
]

[dependency-groups]
dev = []
